+++ apps/codra-demo-video/src/root.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Composition, registerRoot } from "remotion";
-import { CodraDemo } from "./video";
+import { CodraHarnessDemo } from "./video";
 
 export const RemotionRoot = () => {
   return (
     <>
-      <Composition id="CodraDemo" component={CodraDemo} durationInFrames={48 * 30} fps={30} width={1920} height={1080} />
+      <Composition id="CodraHarnessDemo" component={CodraHarnessDemo} durationInFrames={54 * 6} fps={6} width={1920} height={1080} />
     </>
   );
 };